+++ PopupTabLayout/PopupTabLayout.js
@@ -9,6 +9,7 @@
  */
 
 import kind from '@enact/core/kind';
+import {cap} from '@enact/core/util';
 import Spotlight from '@enact/spotlight';
 import PropTypes from 'prop-types';
 import React from 'react';
@@ -194,6 +195,20 @@ const PopupTabLayoutBase = kind({
 		position: PropTypes.oneOf(['left']),
 
 		/**
+		 * Scrim type.
+		 *
+		 * * Values: `'transparent'`, `'translucent'`, or `'none'`.
+		 *
+		 * `'none'` is not compatible with `spotlightRestrict` of `'self-only'`, use a transparent scrim
+		 * to prevent mouse focus when using popup.
+		 *
+		 * @type {('transparent'|'translucent'|'none')}
+		 * @default 'translucent'
+		 * @public
+		 */
+		scrimType: PropTypes.oneOf(['transparent', 'translucent', 'none']),
+
+		/**
 		 * The container id for {@link spotlight/Spotlight}.
 		 *
 		 * @type {String}
@@ -227,12 +242,17 @@ const PopupTabLayoutBase = kind({
 			}
 		},
 		orientation: 'vertical',
-		position: 'left'
+		position: 'left',
+		scrimType: 'translucent'
 	},
 
 	styles: {
 		css,
 		className: 'popupTabLayout'
+	},
+
+	computed: {
+		className: ({scrimType, styler}) => styler.append(`scrim${cap(scrimType)}`)
 	},
 
 	render: ({children, ...rest}) => {
@@ -314,9 +334,9 @@ const PopupTabLayout = PopupTabLayoutDecorator(PopupTabLayoutBase);
  * A shortcut to access {@link sandstone/PopupTabLayout.Tab}
  *
  * @name Tab
+ * @type {sandstone/PopupTabLayout.Tab}
  * @memberof sandstone/PopupTabLayout.PopupTabLayout
  * @extends sandstone/TabLayout.Tab
- * @ui
  */
 PopupTabLayout.Tab = Tab;